requirements.txt
django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
django-cors-headers==4.6.0
django-environ==0.12.0
pytest-django==4.9.0
black==25.1.0
whitenoise==6.8.2
